requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
numpy==1.16.2
pandas==0.24.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
